[create forum v2] move row component to separate file
