+++ src/components/forums/forum/TopicListRow.tsx
@@ -1,0 +1,118 @@
+import { maxBy, isNil } from 'lodash';
+import { useMemo, useCallback } from 'react';
+import Markdown from 'markdown-to-jsx';
+import Jdenticon from 'react-jdenticon';
+import { ForumPost } from '@usedispatch/client';
+
+import { Link } from '../../../components/common';
+
+import { ForumData } from '../../../utils/hooks';
+import { getIdentity } from '../../../utils/identity';
+import { isSuccess } from '../../../utils/loading';
+import { newPublicKey } from '../../../utils/postbox/validateNewPublicKey';
+import { selectForumPosts, selectRepliesFromPosts } from '../../../utils/posts';
+import { usePath } from '../../../contexts/DispatchProvider';
+
+interface TopicListRowProps {
+  topic: ForumPost;
+  forumData: ForumData;
+}
+
+export function TopicListRow(props: TopicListRowProps): JSX.Element {
+  const { topic, forumData } = props;
+  const { buildTopicPath } = usePath();
+  const topicPath = buildTopicPath(
+    forumData.collectionId.toBase58(),
+    topic.postId,
+  );
+
+  const replies: ForumPost[] = useMemo(() => {
+    return selectForumPosts(selectRepliesFromPosts(forumData.posts, topic));
+  }, [forumData]);
+
+  const activtyDate = useCallback((posts: ForumPost[]) => {
+    const dates = posts.map(({ data }) => data.ts);
+    const mostRecentDate = maxBy(dates, (date) => date.getTime());
+    if (!isNil(mostRecentDate)) {
+      const format = (
+        mostRecentDate.getFullYear() !== new Date().getFullYear()
+          ? {
+            year: 'numeric',
+            month: 'short',
+            day: 'numeric',
+          }
+          : {
+            year: undefined,
+            month: 'short',
+            day: 'numeric',
+          }
+      ) as Intl.DateTimeFormatOptions;
+
+      return mostRecentDate.toLocaleDateString(undefined, { ...format });
+    } else {
+      return '-';
+    }
+  }, []);
+
+  const icons = useCallback((posts: ForumPost[]) => {
+    if (posts.length > 0) {
+      const ids = posts.slice(0, 8).map((p) => p.poster.toBase58());
+
+      return (
+          <div className="rowContentIcons">
+            {ids.map((id, index) => {
+              // TODO(andrew) better error-checking on this
+              // newPublicKey call?
+              const pkey = newPublicKey(id);
+              if (isSuccess(pkey)) {
+                const identity = getIdentity(pkey);
+                return (
+                  <div key={index} className="icon">
+                    { (identity != null)
+                      ? <img
+                        src={identity.profilePicture.href}
+                        style={{ borderRadius: '50%' }}
+                      />
+                      : <Jdenticon value={id} alt="posterID" />
+                    }
+                  </div>
+                );
+              } else {
+                return null;
+              }
+            })}
+            {posts.length > 8 ? '...' : null}
+          </div>
+      );
+    } else {
+      return '-';
+    }
+  }, []);
+
+  return (
+      <tr className="row ">
+        <th className="rowSubj">
+          <Link className="" href={topicPath}>
+            <div className="textBox">
+              <Markdown>{topic.data.subj ?? ''}</Markdown>
+            </div>
+          </Link>
+        </th>
+        <td className="rowIconReplies">
+          <Link className="" href={topicPath}>
+            {icons(replies)}
+          </Link>
+        </td>
+        <td className="rowAmountReplies">
+          <Link className="" href={topicPath}>
+            {replies.length}
+          </Link>
+        </td>
+        <td className="rowDate">
+          <Link className="" href={topicPath}>
+            {activtyDate(replies)}
+          </Link>
+        </td>
+      </tr>
+  );
+}
