+++ src/components/forums/index.tsx
@@ -1,21 +1,22 @@
-export * from "./ConnectionAlert";
-export * from "./Notification";
-export * from "./PoweredByDispatch";
+export * from './ConnectionAlert';
+export * from './Notification';
+export * from './PoweredByDispatch';
 
-export * from "./forum/CreateForum";
-export * from "./forum/EditForum";
-export * from "./forum/ForumContent";
-export * from "./forum/ManageModerators";
-export * from "./forum/ManageOwners";
-export * from "./forum/TopicList";
-export * from "./forum/UploadForumBanner";
+export * from './forum/CreateForum';
+export * from './forum/EditForum';
+export * from './forum/ForumContent';
+export * from './forum/ManageModerators';
+export * from './forum/ManageOwners';
+export * from './forum/TopicList';
+export * from './forum/TopicListRow';
+export * from './forum/UploadForumBanner';
 
-export * from "./topic/CreatePost";
-export * from "./topic/EditPost";
-export * from "./topic/GiveAward";
-export * from "./topic/PostContent";
-export * from "./topic/PostsList";
-export * from "./topic/PostReplies";
-export * from "./topic/RoleLabel";
-export * from "./topic/TopicContent";
-export * from "./topic/Votes";
+export * from './topic/CreatePost';
+export * from './topic/EditPost';
+export * from './topic/GiveAward';
+export * from './topic/PostContent';
+export * from './topic/PostsList';
+export * from './topic/PostReplies';
+export * from './topic/RoleLabel';
+export * from './topic/TopicContent';
+export * from './topic/Votes';
